# backend/openapi.generated.yml
openapi: 3.0.3
info:
  title: Activity Tracking API
  version: 1.0.0
  description: Enterprise activity tracking platform API
paths:
...
  /api/companies/{id}/:
    get:
      operationId: api_companies_retrieve
      description: Get company details
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - api
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
          description: ''
    put:
      operationId: api_companies_update
      description: Update company details
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyUpdate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CompanyUpdate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CompanyUpdate'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
          description: ''
    patch:
      operationId: api_companies_partial_update
      description: Partially update company details
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedCompanyUpdate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedCompanyUpdate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedCompanyUpdate'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
          description: ''
...
components:
  schemas:
...
    Company:
      type: object
      description: Serializer for Company model.
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 255
        slug:
          type: string
          maxLength: 255
          pattern: ^[-a-zA-Z0-9_]+$
        logo:
          type: string
          format: uri
          nullable: true
        description:
          type: string
          nullable: true
        website:
          type: string
          format: uri
          nullable: true
          maxLength: 200
        contact_email:
          type: string
          format: email
          maxLength: 254
        contact_phone:
          type: string
          maxLength: 30
        address:
          type: string
          maxLength: 255
        timezone:
          type: string
          maxLength: 50
        language:
          type: string
          maxLength: 10
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - id
      - name
      - slug
      - updated_at
...
    CompanyUpdate:
      type: object
      description: Serializer for updating a company.
      properties:
        name:
          type: string
          maxLength: 255
        logo:
          type: string
          format: uri
          nullable: true
        description:
          type: string
          nullable: true
        website:
          type: string
          format: uri
          nullable: true
          maxLength: 200
        contact_email:
          type: string
          format: email
          maxLength: 254
        contact_phone:
          type: string
          maxLength: 30
        address:
          type: string
          maxLength: 255
        timezone:
          type: string
          maxLength: 50
        language:
          type: string
          maxLength: 10
        is_active:
          type: boolean
      required:
      - name
...
    PatchedCompanyUpdate:
      type: object
      description: Serializer for updating a company.
      properties:
        name:
          type: string
          maxLength: 255
        logo:
          type: string
          format: uri
          nullable: true
        description:
          type: string
          nullable: true
        website:
          type: string
          format: uri
          nullable: true
          maxLength: 200
        contact_email:
          type: string
          format: email
          maxLength: 254
        contact_phone:
          type: string
          maxLength: 30
        address:
          type: string
          maxLength: 255
        timezone:
          type: string
          maxLength: 50
        language:
          type: string
          maxLength: 10
        is_active:
          type: boolean
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT